refactor: create separate AddDeckDropdown component file
